requirements.txt
streamlit==1.28.0
pandas==2.1.0
openpyxl==3.1.2
langgraph==0.0.26
langchain-core==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0